+++ addons/hr_timesheet/__init__.py
@@ -21,6 +21,7 @@
 
 import hr_timesheet
 import wizard
+import report
 
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
