+++ addons/hr_timesheet/report/__init__.py
@@ -1,0 +1,1 @@
+import hr_timesheet_report
