+++ addons/hr_timesheet/report/hr_timesheet_report_view.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <record id="view_hr_timesheet_report_graph" model="ir.ui.view">
+            <field name="name">hr.timesheet.report.graph</field>
+            <field name="model">hr.timesheet.report</field>
+            <field name="arch" type="xml">
+                <graph string="Timesheet" type="pivot">
+                    <field name="user_id" type="row"/>
+                    <field name="date" interval="day" type="col"/>
+                    <field name="quantity" type="measure"/>
+                    <field name="cost" type="measure"/>
+                </graph>
+            </field>
+        </record>
+
+        <record id="view_hr_timesheet_report_search" model="ir.ui.view">
+            <field name="name">hr.timesheet.report.search</field>
+            <field name="model">hr.timesheet.report</field>
+            <field name="arch" type="xml">
+                <search string="Timesheet">
+                    <field name="account_id"/>
+                    <field name="user_id"/>
+                    <filter string="This Month" name="month" icon="terp-go-year" domain="[('date','&lt;=', time.strftime('%Y-%m-%d')),('date','&gt;=',time.strftime('%Y-%m-01'))]" help="month"/>
+                    <group expand="0" string="Extended Filters...">
+                        <field name="general_account_id"/>
+                        <field name="product_id"/>
+                        <field name="journal_id"/>
+                        <field name="company_id" groups="base.group_multi_company"/>
+                        <field name="date"/>
+                    </group>
+                    <group expand="1" string="Group By">
+                        <filter string="User" name="group_user_id" icon="terp-personal" context="{'group_by':'user_id'}"/>
+                        <filter string="Product" name="product" icon="terp-accessories-archiver" context="{'group_by':'product_id'}"/>
+                        <filter string="Analytic Account" name="analytic_account" icon="terp-folder-green" context="{'group_by':'account_id'}"/>
+                        <filter string="General Account" name="general_account" icon="terp-folder-orange" context="{'group_by':'general_account_id'}"/>
+                        <filter string="Company" icon="terp-go-home" context="{'group_by':'company_id'}" groups="base.group_multi_company"/>
+                        <filter string="Day" icon="terp-go-today" context="{'group_by':'day'}" help="Group by day of date"/>
+                        <filter string="Month" icon="terp-go-month" context="{'group_by':'month'}" help="Group by month of date"/>
+                        <filter string="Year" icon="terp-go-year" context="{'group_by':'year'}" help="Group by year of date"/>
+                    </group>
+                </search>
+            </field>
+        </record>
+
+        <record id="action_hr_timesheet_report_stat_all" model="ir.actions.act_window">
+            <field name="name">Timesheet Analysis</field>
+            <field name="res_model">hr.timesheet.report</field>
+            <field name="view_type">form</field>
+            <field name="view_mode">graph</field>
+            <field name="context">{'search_default_year':1,'search_default_month':1,'search_default_group_user_id':1,'group_by_no_leaf':1,'group_by':[]}</field>
+        </record>
+
+        <menuitem action="action_hr_timesheet_report_stat_all"  id="menu_hr_timesheet_report_all"
+            parent="hr.menu_hr_reporting" groups="base.group_hr_manager" sequence="3"/>
+
+    </data>
+</openerp>
