+++ addons/hr_timesheet_sheet/__openerp__.py
@@ -53,7 +53,6 @@ The validation can be configured in the company:
         'hr_timesheet_sheet_view.xml',
         'hr_timesheet_workflow.xml',
         'report/hr_timesheet_report_view.xml',
-        'report/timesheet_report_view.xml',
         'wizard/hr_timesheet_current_view.xml',
         'hr_timesheet_sheet_data.xml',
         'res_config_view.xml',
