+++ addons/hr_timesheet_sheet/report/__init__.py
@@ -19,7 +19,7 @@
 #
 ##############################################################################
 
-import timesheet_report
 import hr_timesheet_report
+
 # vim:expandtab:smartindent:tabstop=4:softtabstop=4:shiftwidth=4:
