+++ addons/hr_timesheet_sheet/report/hr_timesheet_report_view.xml
@@ -1,65 +1,49 @@
 <?xml version="1.0" encoding="utf-8"?>
 <openerp>
     <data>
-        <record id="view_hr_timesheet_report_graph" model="ir.ui.view">
-            <field name="name">hr.timesheet.report.graph</field>
+        <record id="view_timesheet_report_graph" model="ir.ui.view">
+            <field name="name">timesheet.report.graph</field>
             <field name="model">hr.timesheet.report</field>
+            <field name="inherit_id" ref="hr_timesheet.view_hr_timesheet_report_graph"/>
             <field name="arch" type="xml">
-                <graph string="Timesheet" type="pivot">
-                    <field name="user_id" type="row"/>
-                    <field name="date" interval="day" type="col"/>
-                    <field name="quantity" type="measure"/>
-                    <field name="cost" type="measure"/>
-                </graph>
-            </field>
-        </record>
-        <record id="view_hr_timesheet_report_search" model="ir.ui.view">
-            <field name="name">hr.timesheet.report.search</field>
-            <field name="model">hr.timesheet.report</field>
-            <field name="arch" type="xml">
-                <search string="Timesheet">
-                    <filter icon="terp-go-year" string="This Month" name="month" domain="[('date','&lt;=', time.strftime('%Y-%m-%d')),('date','&gt;=',time.strftime('%Y-%m-01'))]" help="month"/>
-                    <field name="account_id" groups="analytic.group_analytic_accounting"/>
-                    <field name="user_id"/>
-                    <group expand="0" string="Extended Filters...">
-                        <field name="general_account_id"/>
-                        <field name="product_id"/>
-                        <field name="journal_id"/>
-                        <field name="company_id" groups="base.group_multi_company"/>
-                        <field name="date" />
-                    </group>
-                    <group expand="1" string="Group By">
-                        <filter string="User" name="group_user_id" icon="terp-personal" context="{'group_by':'user_id'}"/>
-                        <filter string="Product" icon="terp-accessories-archiver" context="{'group_by':'product_id'}"/>
-                        <filter string="Analytic Account" icon="terp-folder-green" context="{'group_by':'account_id'}" groups="analytic.group_analytic_accounting"/>
-                        <filter string="General Account" icon="terp-folder-orange" context="{'group_by':'general_account_id'}"/>
-                        <filter string="Company" icon="terp-go-home" context="{'group_by':'company_id'}" groups="base.group_multi_company"/>
-                        <filter string="Date (day)" context="{'group_by':'date:day'}" help="Group by day of date"/>
-                        <filter string="Date (month)" context="{'group_by':'date:month'}" help="Group by month of date"/>
-                        <filter string="Date (year)" context="{'group_by':'date:year'}" help="Group by year of date"/>
-                    </group>
-                </search>
+                <field name="user_id" position="after">
+                    <field name="state" type="row"/>
+                </field>
+                <field name="cost" position="after">
+                    <field name="total_diff" type="measure"/>
+                    <field name="total_timesheet" type="measure"/>
+                    <field name="total_attendance" type="measure"/>
+                </field>
             </field>
         </record>
 
-        <record id="action_hr_timesheet_report_stat_all" model="ir.actions.act_window">
-            <field name="name">Timesheet Analysis</field>
-            <field name="res_model">hr.timesheet.report</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">graph</field>
-            <field name="context">{'search_default_month':1,'search_default_group_user_id':1,'group_by_no_leaf':1,'group_by':[]}</field>
-            <field name="help" type="html">
-              <p>
-                This report performs analysis on timesheets created by your
-                human resources in the system. It allows you to have a full
-                overview of  entries done by your employees. You can group them
-                by specific selection criteria thanks to the search tool.
-              </p>
+        <record id="view_timesheet_report_search" model="ir.ui.view">
+            <field name="name">timesheet.report.search</field>
+            <field name="model">hr.timesheet.report</field>
+            <field name="inherit_id" ref="hr_timesheet.view_hr_timesheet_report_search"/>
+            <field name="arch" type="xml">
+                <field name="account_id"  position="before">
+                    <filter icon="terp-document-new" string="Draft" domain="[('state','=','draft')]"/>
+                    <filter icon="terp-camera_test" string="Confirmed" domain="[('state','=','confirm')]"/>
+                    <filter icon="terp-dialog-close" string="Done" domain="[('state','=','done')]"/>
+                </field>
+                <field name="general_account_id" position="after">
+                    <field name="to_invoice"/>
+                    <field name="department_id"/>
+                    <field name="date_to"/>
+                    <field name="date_from"/>
+                </field>
+                <filter name="product" position="before">
+                    <filter string="Department" icon="terp-personal+" context="{'group_by':'department_id'}"/>
+                </filter>
+                <filter name="analytic_account" position="before">
+                    <filter string="Type of Invoicing" icon="terp-stock_symbol-selection" context="{'group_by':'to_invoice'}"/>
+                </filter>
+                <filter name="general_account" position="after">
+                    <filter string="Status" icon="terp-stock_effects-object-colorize" context="{'group_by':'state'}"/>
+                </filter>
             </field>
         </record>
-        <menuitem
-            action="action_hr_timesheet_report_stat_all"
-            id="menu_hr_timesheet_report_all"
-            parent="hr.menu_hr_reporting" sequence="3"/>
+
     </data>
 </openerp>
