--- addons/hr_timesheet_sheet/report/timesheet_report_view.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data>
-        <record id="view_timesheet_report_graph" model="ir.ui.view">
-            <field name="name">timesheet.report.graph</field>
-            <field name="model">timesheet.report</field>
-            <field name="arch" type="xml">
-                <graph string="Timesheet" type="pivot" stacked="True">
-                    <field name="user_id" type="row"/>
-                    <field name="state" type="row"/>
-                    <field name="quantity" type="measure"/>
-                    <field name="cost" type="measure"/>
-                    <field name="total_diff" type="measure"/>
-                    <field name="total_timesheet" type="measure"/>
-                    <field name="total_attendance" type="measure"/>
-                </graph>
-            </field>
-        </record>
-        <record id="view_timesheet_report_search" model="ir.ui.view">
-            <field name="name">timesheet.report.search</field>
-            <field name="model">timesheet.report</field>
-            <field name="arch" type="xml">
-                <search string="Timesheet">
-                    <filter icon="terp-document-new" string="Draft" domain="[('state','=','draft')]"/>
-                    <filter icon="terp-camera_test" string="Confirmed" domain="[('state','=','confirm')]"/>
-                    <filter icon="terp-dialog-close" string="Done" domain="[('state','=','done')]"/>
-                    <field name="account_id"/>
-                    <field name="user_id" />
-                    <group expand="0" string="Extended Filters...">
-                        <field name="general_account_id"/>
-                        <field name="to_invoice"/>
-                        <field name="product_id"/>
-                        <field name="department_id"/>
-                        <field name="company_id" groups="base.group_multi_company"/>
-                        <field name="date_to"/>
-                        <field name="date_from"/>
-                    </group>
-                    <group expand="1" string="Group By">
-                        <filter string="User" name="User_id" icon="terp-personal" context="{'group_by':'user_id'}"/>
-                        <filter string="Department" icon="terp-personal+" context="{'group_by':'department_id'}"/>
-                        <filter string="Product" icon="terp-accessories-archiver" context="{'group_by':'product_id'}"/>
-                        <filter string="Type of Invoicing" icon="terp-stock_symbol-selection" context="{'group_by':'to_invoice'}"/>
-                        <filter string="Analytic Account" icon="terp-folder-green" context="{'group_by':'account_id'}"/>
-                        <filter string="General Account" icon="terp-folder-orange" context="{'group_by':'general_account_id'}"/>
-                        <filter string="Status" icon="terp-stock_effects-object-colorize" context="{'group_by':'state'}"/>
-                        <filter string="Company" icon="terp-go-home" context="{'group_by':'company_id'}" groups="base.group_multi_company"/>
-                        <filter string="Day" icon="terp-go-today" context="{'group_by':'day'}" help="Group by day of date"/>
-                        <filter string="Month" icon="terp-go-month" context="{'group_by':'month'}" help="Group by month of date"/>
-                        <filter string="Year" icon="terp-go-year" context="{'group_by':'year'}" help="Group by year of date"/>
-                    </group>
-                </search>
-            </field>
-        </record>
-
-        <record id="action_timesheet_report_stat_all" model="ir.actions.act_window">
-            <field name="name">Timesheet Sheet Analysis</field>
-            <field name="res_model">timesheet.report</field>
-            <field name="view_type">form</field>
-            <field name="view_mode">graph</field>
-            <field name="context">{'search_default_year':1,'search_default_month':1,'search_default_User_id':1,'group_by_no_leaf':1,'group_by':[]}</field>
-            <field name="search_view_id" ref="view_timesheet_report_search"/>
-        </record>
-        <menuitem action="action_timesheet_report_stat_all" id="menu_timesheet_report_all"
-                parent="hr.menu_hr_reporting" sequence="4" groups="base.group_hr_manager"/>
-    </data>
-</openerp>
