+++ addons/hr_timesheet_sheet/security/ir.model.access.csv
@@ -6,4 +6,3 @@ access_hr_timesheet_sheet_sheet_account,hr_timesheet_sheet.sheet.account,model_h
 access_hr_timesheet_report,hr.timesheet.report,model_hr_timesheet_report,base.group_hr_manager,1,1,0,0
 access_hr_analytic_timesheet_system_user,hr.analytic.timesheet.system.user,model_hr_analytic_timesheet,base.group_user,1,0,0,0
 access_hr_timesheet_sheet_sheet_day,hr.timesheet.sheet.sheet.day.user,model_hr_timesheet_sheet_sheet_day,base.group_user,1,1,1,0
-access_timesheet_report,timesheet.report,model_timesheet_report,base.group_hr_manager,1,1,0,0
